adaptive cmdline checking
The recheck now is a bit dynamic.  This is needed as /proc/PID/commandline
is not immediately available.  Hence at the program start we check more often.
For command which alter their environment, this thing should detect the change
now and then, and for long running commands it shall skip all the hassle.